MDEV-25282 Auto-shutdown on idle when socket-activated
Adds max_idle_execution variable with a UINT_MAX default value
that corresponds to the time in seconds under which the
mariadbd executable will run in an idle state.

Under systemd socket activiation this variable will get a 10 minute
default value. This will enable a service to be activated on
connection and fall back to a shutdown state after 10 minutes of
no queries. The systemd socket activation can renable the startup
transparently.

A global variable of server_last_activity is updated on
accepted connections (where max_idle_execution < UINT_MAX)
and when the connection count (standard or extra) is down
to <= 1.

When the main accept loop times out on the max_idle_execution
seconds, and then the server_last_activity is checked along
with if current connection count (standard + extra) is 0
(in case a recently started connection hasn't finished
a query).

To make this neater, in main accept loop moved code to
handle_new_socket_connection that encompases accepting a connection
and the timeout mechanism has been separated too.

Changed when looping though possible connections, loop until the end of the
list and hereby assume two connection can occur on the same poll/select call
and both will be accepted.